6.9.2b
Augmented QP/C ports to Win32 by adding QSpy64/x64 build configurations

+++ ports/win32-qutest/qutest_port.c
@@ -4,8 +4,8 @@
 * @ingroup ports
 * @cond
 ******************************************************************************
-* Last updated for version 6.9.2
-* Last updated on  2021-01-14
+* Last updated for version 6.9.3
+* Last updated on  2021-03-16
 *
 *                    Q u a n t u m  L e a P s
 *                    ------------------------
@@ -136,7 +136,7 @@ uint8_t QS_onStartup(void const *arg) {
     for (rp = result; rp != NULL; rp = rp->ai_next) {
         l_sock = socket(rp->ai_family, rp->ai_socktype, rp->ai_protocol);
         if (l_sock != INVALID_SOCKET) {
-            if (connect(l_sock, rp->ai_addr, rp->ai_addrlen)
+            if (connect(l_sock, rp->ai_addr, (int)rp->ai_addrlen)
                 == SOCKET_ERROR)
             {
                 closesocket(l_sock);
